Changed float to double

+++ drawgraph.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <math.h>
 
-void drawGraph(SDL_Surface *surface, float *coefficients){
+void drawGraph(SDL_Surface *surface, double *coefficients){
 	printf("Printing graph\n");
 	int w = surface->w;
 	int h = surface->h;
@@ -14,7 +14,7 @@ void drawGraph(SDL_Surface *surface, float *coefficients){
 	int B = (int)coefficients[1];
 	int C = (int)coefficients[2];
 	int D = (int)coefficients[3];
-	float S = coefficients[4];
+	double S = coefficients[4];
 	
 	// Draw x line
 	memset(bufc+w*h/2*4, 0xFF, w*4);
@@ -25,11 +25,11 @@ void drawGraph(SDL_Surface *surface, float *coefficients){
 	}
 
 	// Draw pixels
-	float x = -1.0;
+	double x = -1.0;
 	unsigned int x_offset = 0;
 	while(x_offset < w){
 		// 1. Put pixel on screen
-		float y = A*x*x*x + B*x*x + C*x + D;
+		double y = A*x*x*x + B*x*x + C*x + D;
 		unsigned int y_offset = (unsigned int)((y + 1.0)*h/2.0);
 		int bufpos = x_offset + y_offset*w;
 		if(bufpos > 0 && bufpos < w*h){
@@ -37,10 +37,10 @@ void drawGraph(SDL_Surface *surface, float *coefficients){
 		}
 
 		// 2. Calculate dx/dy
-		float derivative = 3*A*x*x + 2*B*x + C;
+		double derivative = 3*A*x*x + 2*B*x + C;
 
 		// 3. Calculate x step
-		float step = S / sqrt(1 + derivative*derivative);
+		double step = S / sqrt(1 + derivative*derivative);
 		x += step;
 		x_offset = (unsigned int)((x + 1.0)*w/2.0);
 	}
